fix: normalise image data in admin library
